+++ src/NP.Common/Extensions/ObjectExtensions.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections.Generic;
 using System;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace NP.Common.Extensions
 {
@@ -75,6 +77,116 @@ namespace NP.Common.Extensions
 		public static string With(this string stringToFormat, params object[] args)
 		{
 			return string.Format(stringToFormat, args);
+		}
+
+		/// <summary>
+		/// Executes the <paramref name="functor"/> if the <paramref name="obj"/> is of <typeparamref name="TObject"/>.
+		/// </summary>
+		/// <typeparam name="TObject"></typeparam>
+		/// <typeparam name="TResult"></typeparam>
+		/// <param name="obj"></param>
+		/// <param name="functor"></param>
+		/// <returns></returns>
+		/// <exception cref="ArgumentNullException"></exception>
+		public static TResult If<TObject,TResult>(this object obj, Func<TObject, TResult> functor)
+		{
+			if (obj == null)
+			{
+				return default(TResult);
+			}
+
+			if (functor == null)
+			{
+				throw new ArgumentNullException(nameof(functor));
+			}
+
+			if (obj is TObject value)
+			{
+				return functor(value);
+			}
+
+			return default(TResult);
+		}
+
+		/// <summary>
+		/// Executes the <paramref name="functor"/> if the <paramref name="obj"/> is of <typeparamref name="TObject"/>.
+		/// </summary>
+		/// <typeparam name="TObject"></typeparam>
+		/// <typeparam name="TResult"></typeparam>
+		/// <param name="obj"></param>
+		/// <param name="functor"></param>
+		/// <param name="cancellationToken"></param>
+		/// <returns></returns>
+		/// <exception cref="ArgumentNullException"></exception>
+		public static async Task<TResult> IfAsync<TObject, TResult>(this object obj, Func<TObject, CancellationToken, Task<TResult>> functor, CancellationToken cancellationToken = default(CancellationToken))
+		{
+			if (obj == null)
+			{
+				return default(TResult);
+			}
+
+			if (functor == null)
+			{
+				throw new ArgumentNullException(nameof(functor));
+			}
+
+			if (obj is TObject value)
+			{
+				return await functor(value, cancellationToken);
+			}
+
+			return default(TResult);
+		}
+
+		/// <summary>
+		/// Executes the <paramref name="action"/> if the <paramref name="obj"/> is of <typeparamref name="TObject"/>.
+		/// </summary>
+		/// <typeparam name="TObject"></typeparam>
+		/// <param name="obj"></param>
+		/// <param name="action"></param>
+		/// <exception cref="ArgumentNullException"></exception>
+		public static void If<TObject>(this object obj, Action<TObject> action)
+		{
+			if (obj == null)
+			{
+				return;
+			}
+
+			if (action == null)
+			{
+				throw new ArgumentNullException(nameof(action));
+			}
+
+			if (obj is TObject value)
+			{
+				action(value);
+			}
+		}
+
+		/// <summary>
+		/// Executes the <paramref name="action"/> if the <paramref name="obj"/> is of <typeparamref name="TObject"/>.
+		/// </summary>
+		/// <typeparam name="TObject"></typeparam>
+		/// <param name="obj"></param>
+		/// <param name="action"></param>
+		/// <param name="cancellationToken"></param>
+		/// <exception cref="ArgumentNullException"></exception>
+		public static async Task IfAsync<TObject>(this object obj, Func<TObject, CancellationToken, Task> action, CancellationToken cancellationToken = default(CancellationToken))
+		{
+			if (obj == null)
+			{
+				return;
+			}
+
+			if (action == null)
+			{
+				throw new ArgumentNullException(nameof(action));
+			}
+
+			if (obj is TObject value)
+			{
+				await action(value, cancellationToken);
+			}
 		}
 
 	}